:recycle: Move confidentiality acknowledgement download button to draft folder.
Renames component to follow consistent naming pattern with other draft action buttons.
